requirements.txt
flask
flask-sqlalchemy
mysql-connector-python
python-dotenv
requests
sqlalchemy
gunicorn
